multi: Move rpctest-based tests to own package
This moves all tests that rely on the rpctest package to their own
package (internal/integration/rpctests) in preparation of switching to
the recently introduced dcrdtest package.
