+++ internal/integration/rpctests/treasury_test.go
@@ -3,7 +3,11 @@
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
-package rpcserver
+// This file is ignored during the regular tests due to the following build tag.
+//go:build rpctest
+// +build rpctest
+
+package rpctests
 
 import (
 	"context"
